IRV-614: inject stats component into config
